Fix API calls to NaClGetExpKidIndex and NaClGetExpParentIndex to properly understand that a signed integer is returned.
BUG= http://code.google.com/p/nativeclient/issues/detail?id=2616
Review URL: https://chromiumcodereview.appspot.com/9476011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7917 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 4359b514837343063fc33920ad86e5cd5339c5b1

+++ src/trusted/validator/x86/ncval_reg_sfi/nc_memory_protect.c
@@ -358,7 +358,7 @@ void NaClMemoryReferenceValidator(NaClValidatorState* state) {
    * by the string instructions Movs and Cmps. In such cases, the
    * ES:RDI must appear before the DS:RSI.
    */
-  uint32_t i;
+  int i;
   NaClInstState* inst_state = state->cur_inst_state;
   NaClExpVector* vector = state->cur_inst_vector;
 
@@ -388,7 +388,7 @@ void NaClMemoryReferenceValidator(NaClValidatorState* state) {
     });
 
   /* Look for assignments on a memory offset. */
-  for (i = 0; i < vector->number_expr_nodes; ++i) {
+  for (i = 0; i < (int) vector->number_expr_nodes; ++i) {
     /* Note: continue (within this loop) is used to indicate
      * that the node describes a safe memory address.
      */
